move transaction label pipe into separate module
