+++ libs/common/src/lib/pipes/index.ts
@@ -1,3 +1,3 @@
 export * from './lto';
 export * from './encode';
-export * from './transaction-label.pipe';
+export * from './transaction-label';
