+++ libs/common/src/lib/pipes/transaction-label/index.ts
@@ -1,0 +1,2 @@
+export * from './transaction-label-pipe.module';
+export * from './transaction-label.pipe';
