+++ libs/common/src/lib/pipes/transaction-label/transaction-label-pipe.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { TransactionLabelPipe } from './transaction-label.pipe';
+
+@NgModule({
+  declarations: [TransactionLabelPipe],
+  exports: [TransactionLabelPipe]
+})
+export class TransactionLabelPipeModule {}
