+++ libs/common/src/lib/pipes/transaction-label/transaction-label.pipe.ts
